Updated UML. Connected Customer with Car for UML. Changed getType function to include car add ons.

+++ src/PaymentBehavior.java
@@ -1,4 +1,4 @@
 public interface PaymentBehavior {
-	public void makePayment(int amount);
+	public void makePayment(double amount);
 	public boolean validate();
 }
